Table formatting improvements for github.
The github markdown renderer seems to *give up* inside html table tags.
So, do even more of the markup in html.  K.

(This is particularly amusing because I'm using atom's preview for
local review before pushing to github, and you'd think since it's a
product of github, maybe... maybe it'd have the same markdown quirks?
Nah.)

Signed-off-by: Eric Myhre <hash@exultant.us>