An exploration of unixfsv2, schemas, intersecting.
(This is opening up a whole new directory tree, the idea of which is
roughly that "it would be nice to collect some experience reports and
user stories, *before* writing final specs"; and we might still want to
be able to easily refer and link to them and accumulate more of them
while we work towards being ready to make final specs proposals.)

(There's correspondingly a series of "in the opinion of this author"
text in this file.  The idea is that other folks might write very
different explorations, and we might merge and retain both of them.)

(I'd actually like to re-issue https://github.com/ipld/specs/issues/130
as a file in this sort of a format.  One can't link to headings in
a github issue; one *can* link to headings in a markdown file, like
the one this diff is adding.  Maybe other organizational formats could
make even better trades, but this deficiency of github issues is
worth noting.)