Brief description of AST as core of selectors.
More expansion and exposition in this area is still needed.

We also need to consider drawing a new picture at this point: this
graphic has done well for ages, but doesn't introduce matchers vs
exploration, which is now an important concept, so we should make new
imagery that does.

This diff is just enough to make things presentable and merge this
branch with the history separation.  More work should come.