Expand library design recommendations doc.
More caveats and purpose preface at the top.

Remark on the possibility of Node/NodeBuilder split in a new section.

Expand upon some of the higher-level functions content with short
examples of possible code.  (I'm a little worried that putting any
syntax at all here might reduce the broadness of audience, but
apparently _something_ is necessary, per review comments in
https://github.com/ipld/specs/pull/241#discussion_r388681785 ...
so maybe this will do it.  I've chosen a syntax that simply doesn't
exist in any language at all (to my knowledge) to make sure everyone
in the world is _equally_ affronted.  Gesundheit.)