Introduce "EDITING.md".
Some content in here covers recommended places to contact the team and
other community members; some is about checklists of concerns we use
when considering changes; some of it is about sheer syntax.
Bit of a grab bag.

The part I'm most interested in getting up and out is the
considerations checklist.  This list gathers a bunch of items we
already informally check through when considering new ideas... and
it seems to be enough of this list is full of recurring (yet also
*terribly* easy to forget on a bad day) concerns, so we might as well
document it -- both for ourselves and for others.